+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.1.8
+ - Added Games/Collections
+   - MegaMan Battle Network Legacy Collection 1 (Thanks Bad-Acetabulum, RealRelativeEase, and Seb)
+   - MegaMan Battle Network Legacy Collection 2 (Thanks Bad-Acetabulum, RealRelativeEase, and Seb)
+ - Fixes
+   - Requires Python 3.9+
+   
 # 0.1.7
  - Added Games/Collections
    - Breakers Collection (Thanks tydog98 and Feilakas)